This is the publish version 1.0

+++ src/make_circuitfile/make_cir.cpp
@@ -36,10 +36,10 @@ int make_cir(double new_value, int ele_num, vector<ele_unit> &element, vector<st
     }        
 
 
-    for (int x = 0; x < element.size(); x++) {
+    for (size_t x = 0; x < element.size(); x++) {
         liness.clear();
         liness.str("");
-        if(x == ele_num){
+        if((int)x == ele_num){
             value = new_value;
         }
         else{
@@ -107,7 +107,7 @@ int make_cir(double new_value, int ele_num, vector<ele_unit> &element, vector<st
         }
         transform(element[x].element.begin(), element[x].element.end(), element[x].element.begin(), ::toupper); 
     }
-    for (int x = 0; x < data_cir_copy.size(); x++) {     //内容を書き換えたdata_cir_copy(バッファ)を全て書く
+    for (size_t x = 0; x < data_cir_copy.size(); x++) {     //内容を書き換えたdata_cir_copy(バッファ)を全て書く
         if(data_cir_copy[x].find(".file") != string::npos || data_cir_copy[x].find(".FILE") != string::npos || data_cir_copy[x].find(".File") != string::npos){
             fpin << ".FILE CIRCUIT" <<  getpid() << ".CSV" << endl;
         }
