+++ src/misc/search_command.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 int search_command(vector<string> arg_arr, const string command){
 
-    for(int i = 0; i < arg_arr.size(); i++){
+    for(size_t i = 0; i < arg_arr.size(); i++){
         if(arg_arr[i] == command){
             return i;
         }
